# PuzzleShop.Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "Autogenerated API info",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Orders/getCart/{userId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "Orders_GetCart",
        "parameters": [
          {
            "type": "integer",
            "name": "userId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Orders/orders/{userId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "Orders_GetAllOrders",
        "parameters": [
          {
            "type": "integer",
            "name": "userId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
    "/api/Orders/confirmOrder/{userId}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "operationId": "Orders_ConfirmOrder",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "userId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          },
          {
            "name": "customerDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CustomerInfoForOrderDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Orders/placeOrder/{userId}/{orderId}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "operationId": "Orders_PlaceOrder",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "userId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "orderId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          },
          {
            "name": "userData",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CustomerInfoForOrderDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Orders/orderItem/{userId}/{itemId}": {
      "delete": {
        "tags": [
          "Orders"
        ],
        "operationId": "Orders_RemoveOrderItem",
        "parameters": [
          {
            "type": "integer",
            "name": "userId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "itemId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CustomerInfoForOrderDto": {
      "type": "object",
      "required": [
        "contactEmail",
        "customerFirstName",
        "customerLastName",
        "address",
        "city",
        "country",
        "postalCode",
        "phone",
        "token"
      ],
      "properties": {
        "contactEmail": {
          "type": "string"
        },
        "customerFirstName": {
          "type": "string"
        },
        "customerLastName": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
  }
}